+++ pkg/globalconfig/doc.go
@@ -1,0 +1,7 @@
+// Copyright (c) 2026 Benjamin Borbe All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+// Package globalconfig loads and validates the user-level dark-factory configuration
+// from ~/.dark-factory/config.yaml.
+package globalconfig